Remove assertions. Not sure if they should be exceptions.
I remember trying to think out the best thing to do here. It might even
have been my suggestion for all I remember, to put the asserts here.

However, this makes tests fail inside the analyzer. Strictly speaking,
asserts should only assert the impossible -- that which, if it occurs,
should fail a test. But I'm trying to test the error condition here,
and there should be a way for such a test to pass.

Now, the assert makes sense from the perspective of "This operation has
no real answer, you shouldn't do this." However, its not simple to
check *when* that's the case. I'd have to either analyze
`attrName.lexeme.split('.')` myself, or I'd have to track when errors
are reported, and somehow correlate that to the property ast.

This makes me think that what we want is an exception....however, I
don't really want an exception either. I want to be able to get an
answer, with some kind of error recovery.

So for now, I'm just removing, but open to the discussion of what
*should be done*.